Separates differentiable exceptions.
