+++ src/utils/exceptions.hpp
@@ -6,28 +6,6 @@
 
 namespace exceptions {
 
-#pragma region Differentiable
-namespace differentiable {
-class BackwardBeforeForwardException : public std::exception {
-  virtual const char *what() const throw() {
-    return "backward was called before forward.";
-  }
-};
-
-class BackwardCalledInEvalModeException : public std::exception {
-  virtual const char *what() const throw() {
-    return "backward cannot be called when set in evaluation mode.";
-  }
-};
-
-class BackwardCalledWithNoInputException : public std::exception {
-  virtual const char *what() const throw() {
-    return "backward cannot be called without a stored input.";
-  }
-};
-} // namespace differentiable
-#pragma endregion Differentiable
-
 #pragma region Eigen
 namespace eigen {
 class InvalidShapeException : public std::exception {
