+++ test/test_activation_functions.cpp
@@ -1,6 +1,6 @@
 #include "activation_functions.hpp"
 #include "exceptions/activation_functions.hpp"
-#include "utils/exceptions.hpp"
+#include "exceptions/differentiable.hpp"
 #include <Eigen/Dense>
 #include <gtest/gtest.h>
