+++ test/test_cross_entropy_loss.cpp
@@ -1,4 +1,5 @@
 #include "cross_entropy_loss.hpp"
+#include "exceptions/differentiable.hpp"
 #include "utils/exceptions.hpp"
 #include <gtest/gtest.h>
